Set working directory with a command-line option for terminator, gnome-terminal
As of Terminator 2.0.1, the working directory Terminator is started from
no longer seems to be used as the working directory for the new shell.
Terminator (and GNOME Terminal) has a command-line option for setting the
working directory, though, so use that.

Keep using the "set working directory, then launch terminal" approach for
other terminal emulators.